ACL: refactoring of the original authentication code.
